Add caching for Go tests and builds
This change adds two type of caching to the Dockerfile:
- A separate layer for go modules so they only need downloading if
  go.mod/go.sum change.
- A cache mount when building or testing the go source to take advantage
  of Go's built-in caching.

With a warm cache, this brought the time to execute `make test` down
from 1m28s to ~15s on a 2017 Macbook Pro (3.5 GHz Dual-Core Intel Core
i7, 16GB RAM).

To ensure the cache mount is used (Docker Engine 19+), DOCKER_BUILDKIT
is set for calls to `docker build`.