Fix a race condition when sending status updates.
When multiple status updates are sent in short succession, the channels receiving updates may block while waiting for the handling goroutine to catch up.

This was causing some updates to be lost and causing the test to time out.

Fixed by sending updates on dedicated goroutines, and timing out on sending an update after 1 second.