Don't require a public image for development
When developing hidden commands, you needed to have a gort image available tagged with the current version. This could cause confusion if you ran locally with Tilt and tried to run "!help",

This change adds an environment variable that allows you to override the image tag used by the default bundle for running the gort binary.

Uses "latest" in the docker-compose file so that you're always going to get your most recent build when testing.