Add the -B flag to set a base directory for client config.
So you can run test instances of Gort without accidentally overwriting any
production profiles you may be using, this change allows the location of the
.gort directory to be overriden with a command line flag.

This change also updates the Tiltfile to take advantage of the change, and
automatically bootstrap the test instance of Gort to save time during development.

The Tilt image for Tilt is also tagged with the version number so pre-release versions
can be tested with the built-in commands.